Fix ft_strlen iterator type

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: maolivei <maolivei@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/31 15:38:23 by maolivei          #+#    #+#             */
-/*   Updated: 2022/03/31 15:40:40 by maolivei         ###   ########.fr       */
+/*   Updated: 2022/04/01 10:59:23 by maolivei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	size;
+	size_t	size;
 
 	size = 0;
 	while (s[size])
